refactor: complete uint64 migration for containerConcurrency
This commit completes the uint64 type migration to eliminate integer overflow
risks throughout the capacity calculation pipeline. Addresses the code review
concern: "Complete the uint64 migration to eliminate all integer conversion warnings"

**Changes**:

1. **API Layer** (revision_lifecycle.go):
   - Changed GetContainerConcurrency() return type: int64 → uint64
   - Added validation: negative values return 0 (defensive)
   - Validates at the API boundary where K8s data enters the system

2. **Function Signatures**:
   - newRevisionThrottler(): int → uint64 parameter
   - pickLBPolicy(): int → uint64 parameter
   - mustCreateRevisionThrottler() test helper: int → uint64

3. **Callers Updated**:
   - throttler.go: Removed int() conversions, use uint64 directly
   - queue.go: Use strconv.FormatUint instead of Itoa
   - pa.go: Safe uint64→int64 conversion (bounded by K8s validation)

4. **Tests Updated**:
   - revision_lifecycle_test.go: Updated expectations to uint64
   - Added test case for negative concurrency validation
   - throttler_test.go: Changed test struct fields to uint64

5. **Bug Fix**:
   - Fixed typo: reg.eventType → req.eventType (line 1049)

**Benefits**:
- No more uint64↔int conversions in hot paths
- Eliminated 3 nolint suppressions
- Type safety enforced at compile time
- Overflow impossible within the activator subsystem

**Safe Conversions Remaining** (at API boundaries only):
- uint64→int64 for PodAutoscalerSpec (K8s API requires int64)
- int→uint64 for array indices (always positive and small)

All tests pass including -race detector.

+++ pkg/activator/net/helpers_test.go
@@ -254,7 +254,7 @@ func addresses(prefix string, n int) []corev1.EndpointAddress {
 // mustCreateRevisionThrottler is a test helper that creates a revisionThrottler and panics on error.
 // Use this in tests where invalid parameters should never occur.
 func mustCreateRevisionThrottler(t *testing.T, revID types.NamespacedName,
-	loadBalancerPolicy *string, containerConcurrency int, proto string,
+	loadBalancerPolicy *string, containerConcurrency uint64, proto string,
 	breakerParams queue.BreakerParams, logger *zap.SugaredLogger,
 ) *revisionThrottler {
 	t.Helper()
